--- rust/cloud-storage/.sqlx/query-2b84b8b3f8249a00c5fdc077b6450804a38baa96682f5c6dc8bb95dddf25d037.json
@@ -1,40 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n            SELECT id, company_id, domain, created_at\n            FROM crm_domains\n            WHERE company_id = $1\n            ORDER BY created_at ASC\n            ",
-  "describe": {
-    "columns": [
-      {
-        "ordinal": 0,
-        "name": "id",
-        "type_info": "Uuid"
-      },
-      {
-        "ordinal": 1,
-        "name": "company_id",
-        "type_info": "Uuid"
-      },
-      {
-        "ordinal": 2,
-        "name": "domain",
-        "type_info": "Text"
-      },
-      {
-        "ordinal": 3,
-        "name": "created_at",
-        "type_info": "Timestamptz"
-      }
-    ],
-    "parameters": {
-      "Left": [
-        "Uuid"
-      ]
-    },
-    "nullable": [
-      false,
-      false,
-      false,
-      false
-    ]
-  },
-  "hash": "2b84b8b3f8249a00c5fdc077b6450804a38baa96682f5c6dc8bb95dddf25d037"
-}
