--- rust/cloud-storage/.sqlx/query-3f084390f0f997762642796059a56b37dc19270a95b8e12e9e86d53824562037.json
@@ -1,53 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n            SELECT c.id, c.team_id, c.email_sync, c.hidden, c.created_at, c.updated_at\n            FROM crm_companies c\n            JOIN crm_domains d ON d.company_id = c.id\n            WHERE c.team_id = $1\n              AND LOWER(d.domain) = $2\n            LIMIT 1\n            ",
-  "describe": {
-    "columns": [
-      {
-        "ordinal": 0,
-        "name": "id",
-        "type_info": "Uuid"
-      },
-      {
-        "ordinal": 1,
-        "name": "team_id",
-        "type_info": "Uuid"
-      },
-      {
-        "ordinal": 2,
-        "name": "email_sync",
-        "type_info": "Bool"
-      },
-      {
-        "ordinal": 3,
-        "name": "hidden",
-        "type_info": "Bool"
-      },
-      {
-        "ordinal": 4,
-        "name": "created_at",
-        "type_info": "Timestamptz"
-      },
-      {
-        "ordinal": 5,
-        "name": "updated_at",
-        "type_info": "Timestamptz"
-      }
-    ],
-    "parameters": {
-      "Left": [
-        "Uuid",
-        "Text"
-      ]
-    },
-    "nullable": [
-      false,
-      false,
-      false,
-      false,
-      false,
-      false
-    ]
-  },
-  "hash": "3f084390f0f997762642796059a56b37dc19270a95b8e12e9e86d53824562037"
-}
