+++ rust/cloud-storage/crm/src/domain/service.rs
@@ -5,7 +5,7 @@ use crate::domain::{
     companies_repo::{CompaniesRepository, CrmCompanyListSort, CrmCompanySoupCursor},
     company_metadata_resolver::CompanyMetadataResolver,
     generic_email_domains::is_generic_email_domain,
-    model::{CrmCompany, CrmCompanyForSoup, CrmContact, CrmError, CrmScopePrecheck},
+    model::{CrmCompanyForSoup, CrmCompanyWithContacts, CrmContact, CrmError, CrmScopePrecheck},
 };
 use chrono::{DateTime, Utc};
 use serde_json::Value;
@@ -13,15 +13,6 @@ use serde_json::Value;
 /// The CrmService exposes operations over CRM records (companies, their
 /// domains and contacts).
 pub trait CrmService: Clone + Send + Sync + 'static {
-    /// Fetches the company for the given team that has `domain` registered
-    /// against it, hydrated with all of the company's domains. Returns
-    /// `Ok(None)` when no company in the team has that domain.
-    fn get_company_by_domain(
-        &self,
-        team_id: &uuid::Uuid,
-        domain: &str,
-    ) -> impl Future<Output = Result<Option<CrmCompany>, CrmError>> + Send;
-
     /// Idempotently records that `email` was seen from the mailbox
     /// identified by `link_id`, for the team `team_id`. Upserts
     /// `crm_companies` (+ `crm_domains`), `crm_contacts`, and
@@ -220,6 +211,19 @@ pub trait CrmService: Clone + Send + Sync + 'static {
         include_hidden: bool,
     ) -> impl Future<Output = Result<Option<CrmContact>, CrmError>> + Send;
 
+    /// Fetch a single CRM company by id, scoped to `team_id`, hydrated
+    /// with all domains, the primary domain's directory metadata, and
+    /// the company's full contact list. `include_hidden = false` hides
+    /// companies whose `hidden` flag is set and filters hidden contacts;
+    /// `true` (admin/owner) reveals every owned company and contact.
+    /// See [`CompaniesRepository::get_company_for_team`].
+    fn get_company_for_team(
+        &self,
+        team_id: &uuid::Uuid,
+        company_id: &uuid::Uuid,
+        include_hidden: bool,
+    ) -> impl Future<Output = Result<Option<CrmCompanyWithContacts>, CrmError>> + Send;
+
     /// Create a comment on a CRM company or contact, optionally as a reply
     /// to an existing thread. See
     /// [`CompaniesRepository::create_crm_comment`]. Authorization (team
@@ -332,17 +336,6 @@ where
     CR: CompaniesRepository,
     R: CompanyMetadataResolver,
 {
-    #[tracing::instrument(skip(self), err)]
-    async fn get_company_by_domain(
-        &self,
-        team_id: &uuid::Uuid,
-        domain: &str,
-    ) -> Result<Option<CrmCompany>, CrmError> {
-        self.companies_repository
-            .get_company_by_domain(team_id, domain)
-            .await
-    }
-
     #[tracing::instrument(skip(self), err)]
     #[allow(clippy::too_many_arguments)]
     async fn populate_contact(
@@ -568,6 +561,18 @@ where
             .await
     }
 
+    #[tracing::instrument(skip(self), err)]
+    async fn get_company_for_team(
+        &self,
+        team_id: &uuid::Uuid,
+        company_id: &uuid::Uuid,
+        include_hidden: bool,
+    ) -> Result<Option<CrmCompanyWithContacts>, CrmError> {
+        self.companies_repository
+            .get_company_for_team(team_id, company_id, include_hidden)
+            .await
+    }
+
     #[tracing::instrument(skip(self, thread_metadata, text, metadata), err)]
     #[allow(clippy::too_many_arguments)]
     async fn create_crm_comment(
@@ -654,14 +659,6 @@ where
 pub struct NoOpCrmService;
 
 impl CrmService for NoOpCrmService {
-    async fn get_company_by_domain(
-        &self,
-        _team_id: &uuid::Uuid,
-        _domain: &str,
-    ) -> Result<Option<CrmCompany>, CrmError> {
-        unimplemented!("NoOpCrmService.get_company_by_domain")
-    }
-
     async fn populate_contact(
         &self,
         _team_id: &uuid::Uuid,
@@ -760,6 +757,15 @@ impl CrmService for NoOpCrmService {
         _contact_id: &uuid::Uuid,
         _include_hidden: bool,
     ) -> Result<Option<CrmContact>, CrmError> {
+        Ok(None)
+    }
+
+    async fn get_company_for_team(
+        &self,
+        _team_id: &uuid::Uuid,
+        _company_id: &uuid::Uuid,
+        _include_hidden: bool,
+    ) -> Result<Option<CrmCompanyWithContacts>, CrmError> {
         Ok(None)
     }
