+++ rust/cloud-storage/crm/src/outbound/companies_repo.rs
@@ -9,8 +9,8 @@ use crate::domain::{
     },
     companies_repo::{CompaniesRepository, CrmCompanyListSort, CrmCompanySoupCursor},
     model::{
-        CrmAddressStatus, CrmCompany, CrmCompanyForSoup, CrmContact, CrmDomain, CrmDomainStatus,
-        CrmError, CrmScopePrecheck, DomainMetadata,
+        CrmAddressStatus, CrmCompany, CrmCompanyForSoup, CrmCompanyWithContacts, CrmContact,
+        CrmDomain, CrmDomainStatus, CrmError, CrmScopePrecheck, DomainMetadata,
     },
 };
 use chrono::{DateTime, Utc};
@@ -69,66 +69,6 @@ impl CompaniesRepositoryImpl {
 }
 
 impl CompaniesRepository for CompaniesRepositoryImpl {
-    #[tracing::instrument(skip(self), err)]
-    async fn get_company_by_domain(
-        &self,
-        team_id: &uuid::Uuid,
-        domain: &str,
-    ) -> Result<Option<CrmCompany>, CrmError> {
-        let normalized_domain = domain.to_ascii_lowercase();
-
-        let company = sqlx::query!(
-            r#"
-            SELECT c.id, c.team_id, c.email_sync, c.hidden, c.created_at, c.updated_at
-            FROM crm_companies c
-            JOIN crm_domains d ON d.company_id = c.id
-            WHERE c.team_id = $1
-              AND LOWER(d.domain) = $2
-            LIMIT 1
-            "#,
-            team_id,
-            normalized_domain,
-        )
-        .fetch_optional(&self.pool)
-        .await
-        .map_err(|e| CrmError::StorageLayerError(e.into()))?;
-
-        let Some(company) = company else {
-            return Ok(None);
-        };
-
-        let domains = sqlx::query!(
-            r#"
-            SELECT id, company_id, domain, created_at
-            FROM crm_domains
-            WHERE company_id = $1
-            ORDER BY created_at ASC
-            "#,
-            company.id,
-        )
-        .fetch_all(&self.pool)
-        .await
-        .map_err(|e| CrmError::StorageLayerError(e.into()))?
-        .into_iter()
-        .map(|row| CrmDomain {
-            id: row.id,
-            company_id: row.company_id,
-            domain: row.domain,
-            created_at: row.created_at,
-        })
-        .collect();
-
-        Ok(Some(CrmCompany {
-            id: company.id,
-            team_id: company.team_id,
-            email_sync: company.email_sync,
-            hidden: company.hidden,
-            created_at: company.created_at,
-            updated_at: company.updated_at,
-            domains,
-        }))
-    }
-
     #[tracing::instrument(skip(self), err)]
     #[allow(clippy::too_many_arguments)]
     async fn populate_contact(
@@ -1332,6 +1272,100 @@ impl CompaniesRepository for CompaniesRepositoryImpl {
             last_interaction: row.last_interaction,
             created_at: row.created_at,
             updated_at: row.updated_at,
+        }))
+    }
+
+    #[tracing::instrument(skip(self), err)]
+    async fn get_company_for_team(
+        &self,
+        team_id: &uuid::Uuid,
+        company_id: &uuid::Uuid,
+        include_hidden: bool,
+    ) -> Result<Option<CrmCompanyWithContacts>, CrmError> {
+        // Mirrors `list_companies_for_soup`'s domain + directory join shape
+        // but scoped to a single id. Domains arrive ordered by their
+        // `created_at` so the first row carries the "primary" domain and
+        // its `crm_domain_directory` row contributes name/description.
+        // First-interaction / last-interaction are aliased into the
+        // CrmCompany.created_at / updated_at slots — same convention as
+        // the soup listing.
+        let rows = sqlx::query!(
+            r#"
+            SELECT
+                c.id                AS "company_id!",
+                c.team_id           AS "company_team_id!",
+                c.email_sync        AS "company_email_sync!",
+                c.hidden            AS "company_hidden!",
+                c.first_interaction AS "company_created_at!",
+                c.last_interaction  AS "company_updated_at!",
+                d.id                AS "domain_id?",
+                d.domain            AS "domain?",
+                d.created_at        AS "domain_created_at?",
+                dd.name             AS "dir_name?",
+                dd.description      AS "dir_description?"
+            FROM crm_companies c
+            LEFT JOIN crm_domains d ON d.company_id = c.id
+            LEFT JOIN crm_domain_directory dd
+                ON LOWER(dd.domain) = LOWER(d.domain)
+            WHERE c.id = $1
+              AND c.team_id = $2
+              AND ($3 OR c.hidden = FALSE)
+            ORDER BY d.created_at ASC NULLS LAST
+            "#,
+            company_id,
+            team_id,
+            include_hidden,
+        )
+        .fetch_all(&self.pool)
+        .await
+        .map_err(|e| CrmError::StorageLayerError(e.into()))?;
+
+        let mut iter = rows.into_iter();
+        let Some(first) = iter.next() else {
+            return Ok(None);
+        };
+
+        // Directory name/description come from the primary domain's row
+        // (which is the first row by `created_at ASC` per the ORDER BY).
+        let name = first.dir_name.clone();
+        let description = first.dir_description.clone();
+
+        let mut company = CrmCompany {
+            id: first.company_id,
+            team_id: first.company_team_id,
+            email_sync: first.company_email_sync,
+            hidden: first.company_hidden,
+            created_at: first.company_created_at,
+            updated_at: first.company_updated_at,
+            domains: Vec::new(),
+        };
+
+        for row in std::iter::once(first).chain(iter) {
+            if let (Some(id), Some(domain), Some(created_at)) =
+                (row.domain_id, row.domain, row.domain_created_at)
+            {
+                company.domains.push(CrmDomain {
+                    id,
+                    company_id: company.id,
+                    domain,
+                    created_at,
+                });
+            }
+        }
+
+        // Fetch contacts via the existing list method so the hidden
+        // semantics stay in one place. Issued only after the company
+        // resolves so cross-team or hidden 404s short-circuit before
+        // a second round trip.
+        let contacts = self
+            .list_contacts_for_company(team_id, company_id, include_hidden)
+            .await?;
+
+        Ok(Some(CrmCompanyWithContacts {
+            company,
+            name,
+            description,
+            contacts,
         }))
     }
