+++ rust/cloud-storage/crm/src/outbound/no_op_resolver.rs
@@ -3,9 +3,9 @@
 //!
 //! Used by binaries that need to construct a [`CrmService`] but never
 //! call [`populate_contact`] — typically HTTP API servers that only
-//! read existing CRM rows via [`get_company_by_domain`]. Wiring in a
-//! real unfurl-backed resolver in those binaries would force them to
-//! carry the reqwest + scraper deps for code paths they never hit.
+//! read existing CRM rows. Wiring in a real unfurl-backed resolver in
+//! those binaries would force them to carry the reqwest + scraper
+//! deps for code paths they never hit.
 //!
 //! If a binary using this resolver ever *does* call `populate_contact`,
 //! the resulting `crm_domain_directory` entries will all have NULL
@@ -17,7 +17,6 @@
 //!
 //! [`CrmService`]: crate::domain::service::CrmService
 //! [`populate_contact`]: crate::domain::service::CrmService::populate_contact
-//! [`get_company_by_domain`]: crate::domain::service::CrmService::get_company_by_domain
 //! [`UnfurlCompanyMetadataResolver`]: crate::outbound::unfurl_resolver::UnfurlCompanyMetadataResolver
 
 use crate::domain::{company_metadata_resolver::CompanyMetadataResolver, model::DomainMetadata};
